+++ src/dao_association/src/models/dao_association.rs
@@ -23,7 +23,7 @@ impl DaoAssociation {
         purpose: String,
         board: Vec<String>,
         members: Vec<String>,
-        created_at: u32,
+        created_at: u64,
     ) -> Self {
         Self {
             parent: Dao::new(
